Merge de cambios echos por Omar 12/10/20

+++ TpIngSoftII/Models/DTOs/PerfilDto.cs
@@ -10,7 +10,7 @@ namespace TpIngSoftII.Models.DTOs
     {
         public int ID { get; set; }
         public int TipoPerfil { get; set; }
-        public float ValorHorario { get; set; }
+        public decimal ValorHorario { get; set; }
 
     }
 }
